Update simpler_connection_alert182.sma
Cvar for  fade-to-black-death to resume on next death, black or reset AFK / auto go-to-spec timer. <mp_autospec_reset 0|1> Makes it more dramatic and realistic to see all black when dying of not want effects. Doubles for a screen saver. In fact that is what it is. Screen saver for AFK gamers to go easy on monitors or more competitive situation. If program of_spectate is being used it will sent to spec or screensaver <mp_autospec # sec >.