Update hook.sma
PENGUIN ROPE HOOK. USE AT YOUR OWN RISK! Stable. Have not put in the code to auto switch instant hook with sv_hookhead type yet. 
The physics are too in the making. In a nutshell we have a bio gravity gun with a grenade on it's back and can make rope for climbing or entangling.
If you are going to use env_rope it is invisible past sys_ticrate 60.