Update hook.sma
Spawning trigger_pushes albeit invisible( running as designed ) - need to pick a model. Wanting it working first. Barnacles should stay put. Aim carefully!
Snarks and Penguins get shot like a gib shooter or as fast as one can +hook emit them. This time when rope died it took the model with it on env_rope.
That's fine still need to refine rope so I can port it to crossbow.