Update of_spectate.sma
That last script was a comedy. This is refined. Later will exploit the feats from that and make killing spec like a good car race in real life a possibility instead of a shocker. Also with getting shot by bots' displacer being thrown into 3rd person, 150 fov was a riot but like might not still be around on this smoother iteration. Although missed both will be back next time around with controls.