Create changemap_trap.sma
Simply attempts to reload server when educated guess is made when somebody is stuck in the void during mapchange where they otherwise have to exit the client, double-click the server on server list, ctrl+tab, etc to try to resolve a blank screen and everybody else is playing. Client in said VOID does not show in natives such as, get_players. Reloading map fixes this. The version here simply reloads server as it is less natives to get the mapname also for now.