Update hook.sma
Fix hook head physics and catch a train. The hook was passing through 1 window then breaking the next. While that is a sweet 'feat' to apply later that is not usefriendly as it stands without player setting it to such a stealthy mode. Door are better breakables still are no_trigger or something. Needs more time.