Update clientemp.sma
Worked fined with 2 people same IP connected same time. Need a lot more players to test! Added more debugging data as nothing was happening earlier. Some cities and regions came up blank which did not make sense I did manually tests and can get the data and can pull it down 1 try explode array etc. Will need look at it again. Bigger cities have been captured so it is not a matter of size or if the buffer contains it or not.