Update gravity_gun.sma
Plugin was too noisy making sounds every round start and anytime CurWeapon was involved or similar, touching the use button, NO GRAVITY GUN IN HAND MIND YOU ANY OF THOSE TIMES, spawn code was wack, so was the task that froze servers, matched the number of objects to be spawned CVAR with the task instead of infinite loop server wedge amateur code. Adding a clip, have no cleaned up comments to make the damn plugin work regardless of the outcome of the debris plugin attempts to spawn. The gun should work no matter what or a CVAR instead if somebody really wants to disable the whole plugin over something so ridiculous. Looking at past comment about the chicken from Italy. Need to make this gat like the original as it would be pointless otherwise. Not focus on spawn new junk as much as using existing map props. Rip monster_generic and toss, etc.