Update proxysnort.sma
When I thought it was slick to stop downloaders well the code needs to be different for that further otherwise it just ruins everything. Also buffersize needed increased to get risk and type of proxy or vpn, should end run-time error. If so now can dev further to make this less touchy if admin wants and leave VPN users alone like vendor recommends in 'MOST" cases.