Update proxysnort.sma
Buffer evidently grew too large to support nothing really. 2 VPN users showed up on Dashboard yet did not touch the treatment function, sooo since that was only change besides a log to when they do touch the treatment Fn and there is a need for larger buffer, I reshimmed it and erased the VPN entries on my local cache to wait for them to reconnect and observe.