Update hook.sma
Got the displacer ball hook head to go. Fixed issue when hook failed. Added debugging CVAR as to not flood console with prints unless that's what you want. Made specific changes on how the penguins are handled because once they explode they are the hook and you no longer have the hook. Also if bots are nearby, JK_BOTTI dev bots when the penguin hook explodes those bots are known to crash entire server as their pointer is removed wrong. I do not have the time to work such AI as it is not revenue generating nor worth the effort, yet. Otherwise that would be self-waypointing CS bots which I do not care about that game in well over 20 years so let them suffer.