Create arabstreets_radio.sma
Originally just wanted the sound to stop when radio was shot on map since it was bugged and never stops. Then decided to add armor gibs and armor on table when radio is shot as well as add speaker ents on all the cameras. There the garage door, I did not like it open and opened by touching. Have to shoot the button. Also the 1hp breakables such as stained glass window and hostage window I made impenetrable. Replaced redundant music that does not stop when radio is destroyed with sheep sounds.