Update traffic_id.sma
While initial intent was to make a dinky no nonsense plugin time causes these expansions to now include country tags on deaths as to merge other much larger work from a geo api that just happened to use the data on death to illustrate. The currency part of that geo was only thing I really liked , released it day of economic sanctions to show how bad the drops were. The api dropped having the currency rate as the feed got inaccurate.