Update and rename scripting/cstrike/respawn1.1.sma to scripting/cstrike/respawn/respawn1.2.sma
Made unsticking into a native for other plugins to share instead of pose risk of several plugins all using unsticking code haphazardly in piece meal fashion leading to server overhead and instability. Lots of refinements. The Alpha code still makes me laugh and is a lot of fun to use.