Update hook.sma
Fix obvious typo on the cstrike side for Cs maps. Make the painted ents temporary then revert. Schlepped feat over to glock attack2 due to ML players not understanding how to bind, by and large these days. This version  of hook does not sport it's own howto. Switched to a bind on a cvar as I did this which breaks Amxx 182 support for now. Found an opportunity with the beam message kill plan.