Update c4_exp.sma
Correct knife being mistaken for C4 with my plant script. Also expanding script since when T get lots of frags C4 will blow in their face. Want to reuse that for TFK players etc. Added a mini think with c4 details to admins as it ticks so I can work on this progressively.