Update spinx_missiles_3.8.7.sma
Remake parachute seekers due to engine limits putting info sets on the chute think each player. Finds an ent and sees if any of the alive players own it then check distance and fire. Tested. Not as effective but was thinking about switching it with heat seekers or just reorganize it again with low, med, hi accuracy levels/ price sets later since that is the net result. The player name heat seekers were poor compared to the ent 'chute' finders. I guess that is traditional. Now the ent seekers are dumbed up.