Update hook.sma
Since anniversary update caching got screwed over and had to move the modname check to set bool to the precache. Also pragma the hell out of it so we can have more fun with the special hook heads this plugin has to offer. Updated Satchel disarmament also so people have a chance to defend themselves from this chaos.