Add files via upload
Test and update per findings. Announce how many bots were purchased and used that variable to do some C4 work to mitigate random bug when client gets the C4 but can't plant it unless it is dropped first at site. Player drops it if bot was controlled previous round for now. A number of players don't want the responsibilty or for bot to have it already.