Update mapchooser.sma
As Dennis used to say it isn't the code that is interesting but rather the net result. Attempting to make mapchooser to not ever go stale when people connect to otherwise empty server and cannot rock the vote because map was already chosen when nobody was online. This is even worse on Counter-Strike. Join adds mp_timelimit time and never a vote end of game.