Update adminvote.sma
Might be consequence of testing first HL mod OP4 without new colored menus. Had to revert and merge idea to mitigate map name being shrunk when casting votes. It is unavoidable for now. Could make a HUD version of vote with CVAR to toggle the old way or this new one.