Create aim_nice.sma
Found 3 versions. Apr 20  2022 was first save. Just putting this here. I do not use. Just a test. It is laughable it is so shotty. I guess that is the entertainment value. Planned on putting this in the buyzones with tradional CT to T biases as well as buying upgraded or cheap aim for recreational purposes. This is for educational purposes. Use and view AT YOUR OWN(PWN) RISK! It isn't that reliable and one of the reasons I am sharing. Hardward EOL migrations. See copyleft license at root of my git. https://github.com/djearthquake/amxx/blob/main/LICENSE